Split backup and restore logic for helper into separate files
