+++ helper/helper.go
@@ -3,12 +3,9 @@ package helper
 import (
 	"bufio"
 	"bytes"
-	"compress/gzip"
 	"flag"
 	"fmt"
-	"io"
 	"os"
-	"os/exec"
 	"runtime/debug"
 	"sort"
 	"strconv"
@@ -87,233 +84,6 @@ func InitializeGlobals() {
 		os.Exit(0)
 	}
 	operating.InitializeSystemFunctions()
-}
-
-/*
- * Backup specific functions
- */
-
-func doBackupAgent() {
-	var lastRead uint64
-	var (
-		finalWriter io.Writer
-		gzipWriter  *gzip.Writer
-		bufIoWriter *bufio.Writer
-		writeHandle io.WriteCloser
-		writeCmd    *exec.Cmd
-	)
-	toc := &utils.SegmentTOC{}
-	toc.DataEntries = make(map[uint]utils.SegmentDataEntry, 0)
-
-	oidList := getOidListFromFile()
-
-	currentPipe = fmt.Sprintf("%s_%d", *pipeFile, oidList[0])
-	/*
-	 * It is important that we create the reader before creating the writer
-	 * so that we establish a connection to the first pipe (created by gpbackup)
-	 * and properly clean it up if an error occurs while creating the writer.
-	 */
-	for i, oid := range oidList {
-		if wasTerminated {
-			return
-		}
-		if i < len(oidList)-1 {
-			log(fmt.Sprintf("Creating pipe for oid %d\n", oidList[i+1]))
-			nextPipe = fmt.Sprintf("%s_%d", *pipeFile, oidList[i+1])
-			createPipe(nextPipe)
-		}
-
-		log(fmt.Sprintf("Opening pipe for oid %d\n", oid))
-		reader, readHandle := getBackupPipeReader(currentPipe)
-		if i == 0 {
-			finalWriter, gzipWriter, bufIoWriter, writeHandle, writeCmd = getBackupPipeWriter(*compressionLevel)
-		}
-
-		log(fmt.Sprintf("Backing up table with oid %d\n", oid))
-		numBytes, err := io.Copy(finalWriter, reader)
-		gplog.FatalOnError(err, strings.Trim(errBuf.String(), "\x00"))
-		log(fmt.Sprintf("Read %d bytes\n", numBytes))
-
-		lastProcessed := lastRead + uint64(numBytes)
-		toc.AddSegmentDataEntry(uint(oid), lastRead, lastProcessed)
-		lastRead = lastProcessed
-
-		lastPipe = currentPipe
-		currentPipe = nextPipe
-		_ = readHandle.Close()
-		removeFileIfExists(lastPipe)
-	}
-
-	/*
-	 * The order for flushing and closing the writers below is very specific
-	 * to ensure all data is written to the file and file handles are not leaked.
-	 */
-	if gzipWriter != nil {
-		_ = gzipWriter.Close()
-	}
-	_ = bufIoWriter.Flush()
-	_ = writeHandle.Close()
-	if *pluginConfigFile != "" {
-		/*
-		 * When using a plugin, the agent may take longer to finish than the
-		 * main gpbackup process. We either write the TOC file if the agent finishes
-		 * successfully or write an error file if it has an error after the COPYs have
-		 * finished. We then wait on the gpbackup side until one of those files is
-		 * written to verify the agent completed.
-		 */
-		log("Uploading remaining data to plugin destination")
-		err := writeCmd.Wait()
-		gplog.FatalOnError(err, strings.Trim(errBuf.String(), "\x00"))
-	}
-	toc.WriteToFileAndMakeReadOnly(*tocFile)
-	log("Finished writing segment TOC")
-}
-
-func getBackupPipeReader(currentPipe string) (io.Reader, io.ReadCloser) {
-	readHandle, err := os.OpenFile(currentPipe, os.O_RDONLY, os.ModeNamedPipe)
-	gplog.FatalOnError(err)
-	// This is a workaround for https://github.com/golang/go/issues/24164.
-	// Once this bug is fixed, the call to Fd() can be removed
-	readHandle.Fd()
-	reader := bufio.NewReader(readHandle)
-	return reader, readHandle
-}
-
-func getBackupPipeWriter(compressLevel int) (io.Writer, *gzip.Writer, *bufio.Writer, io.WriteCloser, *exec.Cmd) {
-	var writeHandle io.WriteCloser
-	var err error
-	var writeCmd *exec.Cmd
-	if *pluginConfigFile != "" {
-		writeCmd, writeHandle = startBackupPluginCommand()
-	} else {
-		writeHandle, err = os.Create(*dataFile)
-		gplog.FatalOnError(err)
-	}
-
-	var finalWriter io.Writer
-	var gzipWriter *gzip.Writer
-	bufIoWriter := bufio.NewWriter(writeHandle)
-	finalWriter = bufIoWriter
-	if compressLevel > 0 {
-		gzipWriter, err = gzip.NewWriterLevel(bufIoWriter, compressLevel)
-		gplog.FatalOnError(err)
-		finalWriter = gzipWriter
-	}
-	return finalWriter, gzipWriter, bufIoWriter, writeHandle, writeCmd
-}
-
-func startBackupPluginCommand() (*exec.Cmd, io.WriteCloser) {
-	pluginConfig := utils.ReadPluginConfig(*pluginConfigFile)
-	cmdStr := fmt.Sprintf("%s backup_data %s %s", pluginConfig.ExecutablePath, pluginConfig.ConfigPath, *dataFile)
-	writeCmd := exec.Command("bash", "-c", cmdStr)
-
-	writeHandle, err := writeCmd.StdinPipe()
-	gplog.FatalOnError(err)
-	writeCmd.Stderr = &errBuf
-	err = writeCmd.Start()
-	gplog.FatalOnError(err)
-	return writeCmd, writeHandle
-}
-
-/*
- * Restore specific functions
- */
-
-func doRestoreAgent() {
-	tocEntries := utils.NewSegmentTOC(*tocFile).DataEntries
-	var lastByte uint64
-	oidList := getOidListFromFile()
-
-	currentPipe = fmt.Sprintf("%s_%d", *pipeFile, oidList[0])
-	log(fmt.Sprintf("Opening pipe for oid %d", oidList[0]))
-	/*
-	 * It is important that we create the writer before creating the reader
-	 * so that we establish a connection to the first pipe (created by gprestore)
-	 * and properly clean it up if an error occurs while creating the reader.
-	 */
-	writer, writeHandle = getRestorePipeWriter(currentPipe)
-	reader := getRestorePipeReader()
-
-	for i, oid := range oidList {
-		if wasTerminated {
-			return
-		}
-		log(fmt.Sprintf("Restoring table with oid %d", oid))
-		if i < len(oidList)-1 {
-			nextPipe = fmt.Sprintf("%s_%d", *pipeFile, oidList[i+1])
-			createPipe(nextPipe)
-		} else {
-			nextPipe = ""
-		}
-		start := tocEntries[uint(oid)].StartByte
-		end := tocEntries[uint(oid)].EndByte
-		log(fmt.Sprintf("Start Byte: %d; End Byte: %d; Last Byte: %d", start, end, lastByte))
-		_, err := reader.Discard(int(start - lastByte))
-		gplog.FatalOnError(err)
-		log(fmt.Sprintf("Discarded %d bytes", start-lastByte))
-		bytesRead, err := io.CopyN(writer, reader, int64(end-start))
-		log(fmt.Sprintf("Read %d bytes", bytesRead))
-		gplog.FatalOnError(err, errBuf.String())
-		log(fmt.Sprintf("Closing pipe for oid %d", oid))
-		flushAndCloseRestoreWriter()
-		lastByte = end
-
-		lastPipe = currentPipe
-		currentPipe = nextPipe
-		removeFileIfExists(lastPipe)
-		if currentPipe != "" {
-			log(fmt.Sprintf("Opening pipe for oid %d", oid))
-			writer, writeHandle = getRestorePipeWriter(currentPipe)
-		}
-	}
-}
-
-func getRestorePipeReader() *bufio.Reader {
-	var readHandle io.Reader
-	var err error
-	if *pluginConfigFile != "" {
-		readHandle = startRestorePluginCommand()
-		defer func() {
-			if len(errBuf.String()) != 0 {
-				gplog.Error(errBuf.String())
-			}
-		}()
-	} else {
-		readHandle, err = os.Open(*dataFile)
-		gplog.FatalOnError(err)
-	}
-
-	var bufIoReader *bufio.Reader
-	if strings.HasSuffix(*dataFile, ".gz") {
-		gzipReader, err := gzip.NewReader(readHandle)
-		gplog.FatalOnError(err)
-		bufIoReader = bufio.NewReader(gzipReader)
-	} else {
-		bufIoReader = bufio.NewReader(readHandle)
-	}
-	return bufIoReader
-}
-
-func startRestorePluginCommand() io.Reader {
-	pluginConfig := utils.ReadPluginConfig(*pluginConfigFile)
-	cmdStr := fmt.Sprintf("%s restore_data %s %s", pluginConfig.ExecutablePath, pluginConfig.ConfigPath, *dataFile)
-	cmd := exec.Command("bash", "-c", cmdStr)
-
-	readHandle, err := cmd.StdoutPipe()
-	gplog.FatalOnError(err)
-	cmd.Stderr = &errBuf
-
-	err = cmd.Start()
-	gplog.FatalOnError(err)
-	return readHandle
-
-}
-
-func getRestorePipeWriter(currentPipe string) (*bufio.Writer, *os.File) {
-	fileHandle, err := os.OpenFile(currentPipe, os.O_WRONLY, os.ModeNamedPipe)
-	gplog.FatalOnError(err)
-	pipeWriter := bufio.NewWriter(fileHandle)
-	return pipeWriter, fileHandle
 }
 
 /*
